% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotime_axis.R
\name{geotime_axis}
\alias{geotime_axis}
\title{The geotime_axis function}
\usage{
geotime_axis(
  plot,
  level = "period",
  age_min = 0,
  age_max = 540,
  x_reverse = T,
  y_reverse = F,
  label = "abbr",
  div_y = 10,
  div_x = 10,
  pos = "right"
)
}
\arguments{
\item{plot}{An intiated ggplot object without any theme or axis settings}

\item{level}{Level of the geological interval to be shown on the axis. Default is period. Others available include age, epoch, era, and eon.}

\item{age_min}{The minimum age of the plot (in Ma). Default is 0.}

\item{age_max}{The maximum age of the plot (in Ma). Default is 540.}

\item{x_reverse}{If the x axis should be reversed (older ages to the left)}

\item{label}{If the time scale should be labelled. Options: abbr, full, or none}

\item{div_y}{How many breaks (tickmarks) should be on the y axis. Default is 10.}

\item{div_x}{How many breaks (tickmarks) should be on the x axis. Default is 10.}

\item{pos}{The position of the legend if there is one. Default is to the right. See ggplot2.}

\item{y_reveerse}{If the y axis should be reversed (smaller value to the top)}
}
\description{
This function allows you to do zonal statistics of a point dataset
}
\examples{
plot <- ggplot() + geom_line(aes(...), ...) 
p <- geotime_axis(plot, age_min = 0, age_max = 800,level = "period", label = "abbr", div_x = 10)
}
\keyword{geological}
\keyword{scale}
\keyword{time}
