% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereograph.R
\name{stereograph}
\alias{stereograph}
\title{The stereograph function}
\usage{
stereograph(strike, dip)
}
\arguments{
\item{strike}{Strike}

\item{dip}{Dip}
}
\description{
This function is for plotting the stereograph from a pair of strike/dip data
}
\examples{
stereograph(112, 45)
}
\keyword{and}
\keyword{dip}
\keyword{strike}
