% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{zonal_stats}
\alias{zonal_stats}
\title{The zonal_stats function}
\usage{
zonal_stats(
  data,
  res = 13,
  land_color = "tan",
  marine_color = "light blue",
  color_ramp = c("pink", "red"),
  label = "records",
  level = 10,
  map = T,
  stat = sum,
  ...
)
}
\arguments{
\item{data}{The input dataframe. It should contain 3 columns: lat, lon, and value}

\item{res}{The resolution (grid size) of the zonal statistics. Default is 13. The larger the value, the coarser the grid.}

\item{land_color}{The color for the land area in the map (optional)}

\item{marine_color}{The color for the ocean area in the map (optional)}

\item{color_ramp}{A list of 2 colors indicating the color ramp of the map (for the lowest and highest values. optional)}

\item{label}{The label for the map legend. Default is "records"}

\item{level}{The number of color classes in the map. Default is 10}

\item{map}{Whether to plot the map. Default is TRUE}

\item{stat}{The type of zonal statistics. Default is sum. Other available options are min, max, mean, median, var, and sd}
}
\description{
This function allows you to do zonal statistics of a point dataset
}
\examples{
zonal_stats(data, land_color="tan", marine_color="light blue", color_ramp=c("pink","red"), res= 20, level = 15, label = "value", stat=min)
}
\keyword{statistics}
\keyword{zonal}
